dspy-hitl-metric/pyproject.toml
[project]
name = "dspy-hitl-metric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.6.27",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
]
